Extract SQL handling into pkg postgres
